84: Move newInvoice to batch_create/newInvoice
